update price per day type

+++ rental-car-backend/database/migrations/2024_09_14_220626_create_cars_table.php
@@ -13,19 +13,20 @@ return new class extends Migration
     {
         Schema::create('cars', function (Blueprint $table) {
             $table->id();
-            $table->string('Name');             // Car make, e.g., Toyota
-            $table->string('Model');            // Car model, e.g., Corolla
-            $table->string('Year');             // Year of manufacture
-            $table->string('Doors');            // Number of doors
-            $table->string('AC');               // Air conditioning (Yes/No)
-            $table->string('Transmission');     // Transmission type (Manual/Automatic)
-            $table->string('Fuel');             // Fuel type (Petrol/Diesel)
-            $table->decimal('Price_per_day', 8, 2); // Rental price per day
-            $table->string('Status');           // Car status (Available, Rented, Maintenance)
+            $table->string('Name');
+            $table->string('Model');
+            $table->string('Year');
+            $table->string('Doors');
+            $table->string('AC');
+            $table->string('Transmission');
+            $table->string('Fuel');
+            // Changing Price_per_day to store as an integer (in cents)
+            $table->integer('Price_per_day'); 
+            $table->string('Status');
             $table->timestamps();
         });
-        
     }
+    
 
     /**
      * Reverse the migrations.
